Internal #861: Refactor WindowCustomAggregator
Move the WindowCustomAggregator class to separate files.
